requirements.txt
Flask>=0.11.0
Flask-SocketIO>=2.4.0
gevent>=1.2.2
Pillow>=3.1.2
gps>=3.15
